moved three support functions to attributes.js to tighten core lib readability
